changed the version listener to iomplement ResourceListener
